requirements.txt
pyside6~=6.6.1
qtawesome~=1.3.0
pillow~=10.2.0
africastalking~=1.2.7
cryptography~=41.0.7
setuptools~=68.2.2
barcode~=1.0.4